feat: preemptively rotate 0-conf API WebSocket
