+++ docs/boltz.conf
@@ -332,9 +332,11 @@ feeFloor = 0.1
 # mempool check is used.
 # [liquid.chain.zeroConfTool]
 # endpoint = "https://zeroconf.liquid.example/api/v1/zeroconf"  # or ws(s)://.../ws/v1/zeroconf
-# interval = 100        # HTTP only: poll cadence in ms (default 100)
-# max_retries = 60      # HTTP only: retries before client-side reject (default 60)
-# deadline_secs = 6     # WS only:   per-tx wall-time deadline in seconds (default 6)
+#
+# interval = 100                 # HTTP only: poll cadence in ms (default 100)
+# max_retries = 60               # HTTP only: retries before client-side reject (default 60)
+# deadline_secs = 6              # WS only:   per-tx wall-time deadline in seconds (default 6)
+# rotation_interval_secs = 3300  # WS only: proactively reconnect before server TTL (0 disables)
 
 # =============================================================================
 # EVM Chain Configuration (RSK)
